requirements.txt
# Core Dependencies
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Web UI
streamlit>=1.28.0

# Development Tools
pytest>=7.3.0
pytest-cov>=4.1.0
hypothesis>=6.80.0
mypy>=1.4.0
black>=23.0.0
ruff>=0.1.0

# Optional RL Libraries (commented out by default)
# stable-baselines3>=2.0.0
# torch>=2.0.0
